Coerce Rational to Integer to allow usage of & operator.
I'm not entirely sure how to reproduce this in a minimal way, but sometimes the the expr
(localspc/8 + @state.dirty.length) ends up being an instance of Rational (part of Ruby
stdlib), which does not have the & operator defined. This works around the issue by
forcing the Rational back to an Integer.