Avoid generating labels with '..' in them
The C compiler generates labels using "%x" % string.object_id. If the pointer
for string.object_id begins with the most significant digit set, it looks like
a sign-extended 2's complement number (negative), and gets formatted by ruby as
'..f1412300'. On 32-bit platforms, there is rather high chance of randomly
ending up with a label like 'goto_test_uuid..f1234560:', which is a parse error
for the assembler.

This patch simply takes the absolute value of the object_id to avoid negative
interpretations.